test: v4l2_videodevice: buffer_sharing: Lower resolution to speed up test
Lower the resolution to the smallest vivid supports, 320x180, in order
to speed up the test. This isn't enough in itself as the frame rate also
has to be configured, but the V4L2VideoDevice class doesn't support
frame rate control yet. In the meantime the frame rate can be set
manually with yavta or v4l2-ctl.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Paul Elder <paul.elder@ideasonboard.com>