readme: Convert from Markdown to RST
The Documentation and top-level README both serve to provide information
to new developers and users of libcamera.

Currently our README.md is sparse, and should be expanded with more
useful information such as build requirements as well as compilation
directions.

Any information provided in the README also serves a purpose in the main
documentation and website and could be a cause of duplicated content.

To allow direct integration of the top level README document into our
sphinx documentation (and thus the libcamera.org website) convert the
README.md into ReSTructured text format.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>