libcamera: controls: Add AeLocked control
Add a control to report lock status of the Auto Exposure algorithm. The
idea is that if an AE algorithm is running the control shall be added to
the metadata control list.

If the AE algorithm is locked the value shall be set to true, if it's
converging it shall be set to false. If the AE algorithm is not running
the control shall not be present in the metadata control list.

Signed-off-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>