libcamera: v4l2_controls: Turn V4L2ControlInfoMap into a class
In preparation for extending V4L2ControlInfoMap with control idmap
support, turn it into a real class. Make it inherit from std::map<>
instead of wrapping it to keep the API simple.

V4L2ControlInfoMap is meant to be constructed with a set of control
info, and never modified afterwards. To ensure this, inherit from
std::map<> with private access specifier, and explicitly expose the
std::map<> methods that do not allow insertion or removal of elements. A
custom move assignment operator is implemented to allow efficient
construction of the V4L2ControlInfoMap.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Jacopo Mondi <jacopo@jmondi.org>
Tested-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>
Reviewed-by: Niklas Söderlund <niklas.soderlund@ragnatech.se>